agent/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Birdhouse"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.24",
    "langgraph>=0.3.34",
    "pydantic>=2.11.3",
    "pytest>=8.3.5",
    "python-dotenv",
    "modal",
    "fastapi>=0.111.5",
    "uvicorn",
    "litellm>=1.63.0",
    "langchain_openai"
]

[tool.setuptools.packages.find]
where = ["tutorials/langchain_example"]